Remove `ChatGPTVisionConfig` and refactor configs (#539)
