+++ test/models/test_openai_model.py
@@ -29,7 +29,6 @@ from camel.utils import OpenAITokenCounter
         ModelType.GPT_4,
         ModelType.GPT_4_32K,
         ModelType.GPT_4_TURBO,
-        ModelType.GPT_4_TURBO_VISION,
     ],
 )
 def test_openai_model(model_type):
